Split python code into their own files
